chore: consolidate logger module
- Make one logger module
- Remove ConfigurationError to use it from exceptions
